requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.128.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.46.0
MarkupSafe==3.0.3
murmurhash==1.0.15
numba==0.63.1
numpy==2.3.5
packaging==26.0
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pynndescent==0.6.0
python-multipart==0.0.22
requests==2.32.5
ruff==0.14.14
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
